refactor: constant 파일을 각 도메인에 맞게 분리 및 eslint 규칙 적용
