+++ packages/shared/src/constant/category.type.ts
@@ -1,0 +1,5 @@
+export type Category = {
+  id: number;
+  name: string;
+  parent: Category | undefined;
+};
