+++ packages/shared/src/index.ts
@@ -1,3 +1,7 @@
-export * from "./constant";
+export * from "./constant/category.type";
+export * from "./constant/game.type";
+export * from "./constant/question.type";
+export * from "./constant/user-info.type";
+
 export * from "./api-types";
 export * from "./socket-types";
